Moved cookie to own file, removed transport layer, renamed DatagramTransportLayer to DatagramSession
