+++ rustls/src/lib.rs
@@ -239,7 +239,6 @@ mod client;
 mod key;
 mod bs_debug;
 mod keylog;
-mod transport_layer;
 
 /// Internal classes which may be useful outside the library.
 /// The contents of this section DO NOT form part of the stable interface.
