+++ rustls/src/msgs/enums.rs
@@ -17,7 +17,7 @@ enum_builder! {
         TLSv1_3 => 0x0304,
         DTLSv1_0 => 0xFEFF,
         DTLSv1_2 => 0xFEFD,
-        DTLSv1_3 => 0xFEFB,
+        DTLSv1_3 => 0xFEFB
     }
 }
 
@@ -110,7 +110,6 @@ enum_builder! {
         HelloRequest => 0x00,
         ClientHello => 0x01,
         ServerHello => 0x02,
-        #[cfg(feature="dtls")]
         HelloVerifyRequest => 0x03,
         NewSessionTicket => 0x04,
         EndOfEarlyData => 0x05,
